security with token

+++ planner/routes/events.py
@@ -1,11 +1,11 @@
 from typing import List
 
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Body, HTTPException, status, Depends
 from beanie import PydanticObjectId
 
 from database.connection import Database
 from models.events import Event, EventUpdate
-
+from auth.authenticate import authenticate
 
 
 event_router = APIRouter(tags=["Events"])
@@ -30,13 +30,21 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
     
 
 @event_router.post("/new")
-async def create_event(body: Event = Body(...)) -> dict:
+async def create_event(body: Event = Body(...), user: str = 
+                       Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {"message": "Event created successfully"}
 
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
+async def update_event(id: PydanticObjectId, body: EventUpdate,
+                       user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation is not allowed!")
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException (
@@ -46,12 +54,18 @@ async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
 
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
-    event = await event_database.delete(id)
+async def delete_event(id: PydanticObjectId, user: str = 
+                       Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
     if not event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Event with supplied ID does not exist")
+            detail="Event not found")
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation is not allowed!")
+    event = await event_database.delete(id)
     return {"message": "Event deleted successfully."}
